Check for config changes every second instead of every frame, also more refactor
