+++ Source/MaterialColor/MaterialColor.csproj
@@ -48,6 +48,10 @@
       <HintPath>..\lib\Assembly-CSharp-firstpass.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="Assembly-CSharp-firstpass">
+      <HintPath>..\lib\Assembly-CSharp-firstpass.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="UnityEngine, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
@@ -72,15 +76,17 @@
     <Compile Include="Extensions\ColorExtensions.cs" />
     <Compile Include="Helpers\DebugHelper.cs" />
     <Compile Include="Data\FilterInfo.cs" />
+    <Compile Include="IO\ConfigWatcher.cs" />
+    <Compile Include="Painter.cs" />
     <Compile Include="TextFilter.cs" />
-    <Compile Include="Json\BaseManager.cs" />
-    <Compile Include="Json\ConfiguratorStateManager.cs" />
-    <Compile Include="Json\ElementColorsManager.cs" />
-    <Compile Include="Json\JsonFileLoader.cs" />
-    <Compile Include="Paths.cs" />
+    <Compile Include="IO\Json\BaseManager.cs" />
+    <Compile Include="IO\Json\ConfiguratorStateManager.cs" />
+    <Compile Include="IO\Json\ElementColorsManager.cs" />
+    <Compile Include="IO\Json\JsonFileLoader.cs" />
+    <Compile Include="IO\Paths.cs" />
     <Compile Include="State.cs" />
     <Compile Include="Helpers\ColorHelper.cs" />
-    <Compile Include="Harmony\HarmonyPatches.cs" />
+    <Compile Include="HarmonyPatches.cs" />
     <Compile Include="Helpers\MaterialHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="IDs.cs" />
